Added a common methods file for methods that are common to both omc and imu Added file descriptions
